Remove installation of custom table from VIP

+++ laterpay/uninstall.php
@@ -5,27 +5,34 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
     exit;
 }
 
+if ( defined( 'WPCOM_IS_VIP_ENV' ) && true === WPCOM_IS_VIP_ENV ) {
+    return;
+}
+
 global $wpdb;
+
+// Check for non-VIP env.
 
 $table_terms_price   = $wpdb->prefix . 'laterpay_terms_price';
 $table_history       = $wpdb->prefix . 'laterpay_payment_history';
 $table_post_views    = $wpdb->prefix . 'laterpay_post_views';
 $table_time_passes   = $wpdb->prefix . 'laterpay_passes';
 $table_subscriptions = $wpdb->prefix . 'laterpay_subscriptions';
-$table_postmeta      = $wpdb->postmeta;
-$table_usermeta      = $wpdb->usermeta;
 
 // remove custom tables
 $sql = "
-    DROP TABLE IF EXISTS
-        $table_terms_price,
-        $table_history,
-        $table_post_views,
-        $table_time_passes,
-        $table_subscriptions
-    ;
+DROP TABLE IF EXISTS
+    $table_terms_price,
+    $table_history,
+    $table_post_views,
+    $table_time_passes,
+    $table_subscriptions
+;
 ";
 $wpdb->query( $sql );
+
+$table_postmeta      = $wpdb->postmeta;
+$table_usermeta      = $wpdb->usermeta;
 
 // remove pricing and voting data from wp_postmeta table
 delete_post_meta_by_key( 'laterpay_post_prices' );
